+++ framework/scripts/wazuh-clusterd.py
@@ -7,6 +7,7 @@ try:
     import asynchat
     import socket
     import json
+    import ast
     from distutils.util import strtobool
     from sys import argv, exit, path
     from os.path import dirname
@@ -39,10 +40,18 @@ try:
         myWazuh = Wazuh(get_init=True)
 
         from wazuh.common import *
-        from wazuh.cluster import *
+        from wazuh.cluster.management import *
+        from wazuh.cluster.handler import *
+        from wazuh.cluster.distributed_api import *
+        from wazuh.cluster.protocol_messages import *
+        from wazuh.agent import Agent
         from wazuh.exception import WazuhException
         from wazuh.utils import check_output
         from wazuh.pyDaemonModule import pyDaemon, create_pid, delete_pid
+        import wazuh.syscheck as syscheck
+        import wazuh.rootcheck as rootcheck
+        import wazuh.stats as stats
+        import wazuh.manager as manager
     except Exception as e:
         print("Error importing 'Wazuh' package.\n\n{0}\n".format(e))
         exit()
@@ -92,17 +101,222 @@ class WazuhClusterHandler(asynchat.async_chat):
             res = "Received invalid cluster command {0}".format(self.command[0])
 
         if error == 0:
-            if self.command[0] == 'node':
+            if self.command[0] == list_requests_cluster['node']:
                 res = get_node()
-            elif self.command[0] == 'zip':
+            elif self.command[0] == list_requests_cluster['zip']:
                 zip_bytes = self.f.decrypt(response[common.cluster_sync_msg_size:])
                 res = extract_zip(zip_bytes)
-                self.restart = res['restart']
-            elif self.command[0] == 'ready':
-                # sync_one_node(False, self.addr)
+            elif self.command[0] == list_requests_agents['RESTART_AGENTS']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                if (len(args) == 2):
+                    agents = args[0].split("-")
+                    restart_all = ast.literal_eval(args[1])
+                else:
+                    agents = None
+                    restart_all = ast.literal_eval(args[0])
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = Agent.restart_agents(agents, restart_all, cluster_depth)
+            elif self.command[0] == list_requests_agents['AGENTS_UPGRADE_RESULT']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                try:
+                    agent = args[0]
+                    timeout = args[1]
+                    res = Agent.get_upgrade_result(agent, timeout)
+                except Exception as e:
+                    res = str(e)
+            elif self.command[0] == list_requests_agents['AGENTS_UPGRADE']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                agent_id = args[0]
+                wpk_repo = ast.literal_eval(args[1])
+                version = ast.literal_eval(args[2])
+                force = ast.literal_eval(args[3])
+                chunk_size = ast.literal_eval(args[4])
+                try:
+                    res = Agent.upgrade_agent(agent_id, wpk_repo, version, force, chunk_size)
+                except Exception as e:
+                    res = str(e)
+            elif self.command[0] == list_requests_agents['AGENTS_UPGRADE_CUSTOM']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                agent_id = args[0]
+                file_path = ast.literal_eval(args[1])
+                installer = ast.literal_eval(args[2])
+                try:
+                    res = Agent.upgrade_agent_custom(agent_id, file_path, installer)
+                except Exception as e:
+                    res = str(e)
+            elif self.command[0] == list_requests_syscheck['SYSCHECK_LAST_SCAN']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                agent = args.split(" ")
+                res = syscheck.last_scan(agent[0])
+            elif self.command[0] == list_requests_syscheck['SYSCHECK_RUN']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                if (len(args) == 2):
+                    agents = args[0]
+                    all_agents = ast.literal_eval(args[1])
+                else:
+                    agents = None
+                    all_agents = ast.literal_eval(args[0])
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = syscheck.run(agents, all_agents, cluster_depth)
+            elif self.command[0] == list_requests_syscheck['SYSCHECK_CLEAR']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                if (len(args) == 2):
+                    agents = args[0]
+                    all_agents = ast.literal_eval(args[1])
+                else:
+                    agents = None
+                    all_agents = ast.literal_eval(args[0])
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = syscheck.clear(agents, all_agents, cluster_depth)
+            elif self.command[0] == list_requests_rootcheck['ROOTCHECK_PCI']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                index = 0
+                agents = None
+                if (len(args) == 5):
+                    agents = args[0]
+                    index = index + 1
+                offset = ast.literal_eval(args[index])
+                index = index + 1
+                limit = ast.literal_eval(args[index])
+                index = index + 1
+                sort = ast.literal_eval(args[index])
+                index = index + 1
+                search = ast.literal_eval(args[index])
+                res = args
+                res = rootcheck.get_pci(agents, offset, limit, sort, search)
+            elif self.command[0] == list_requests_rootcheck['ROOTCHECK_CIS']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                index = 0
+                agents = None
+                if (len(args) == 5):
+                    agents = args[0]
+                    index = index + 1
+                offset = ast.literal_eval(args[index])
+                index = index + 1
+                limit = ast.literal_eval(args[index])
+                index = index + 1
+                sort = ast.literal_eval(args[index])
+                index = index + 1
+                search = ast.literal_eval(args[index])
+                res = args
+                res = rootcheck.get_cis(agents, offset, limit, sort, search)
+            elif self.command[0] == list_requests_rootcheck['ROOTCHECK_LAST_SCAN']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                agent = args.split(" ")
+                res = rootcheck.last_scan(agent[0])
+            elif self.command[0] == list_requests_rootcheck['ROOTCHECK_RUN']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                if (len(args) == 2):
+                    agents = args[0]
+                    all_agents = ast.literal_eval(args[1])
+                else:
+                    agents = None
+                    all_agents = ast.literal_eval(args[0])
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = rootcheck.run(agents, all_agents, cluster_depth)
+            elif self.command[0] == list_requests_rootcheck['ROOTCHECK_CLEAR']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                if (len(args) == 2):
+                    agents = args[0]
+                    all_agents = ast.literal_eval(args[1])
+                else:
+                    agents = None
+                    all_agents = ast.literal_eval(args[0])
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = rootcheck.clear(agents, all_agents, cluster_depth)
+            elif self.command[0] == list_requests_managers['MANAGERS_STATUS']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = manager.managers_status(cluster_depth=cluster_depth)
+            elif self.command[0] == list_requests_managers['MANAGERS_LOGS']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                type_log = args[0]
+                category = args[1]
+                months = ast.literal_eval(args[2])
+                offset = ast.literal_eval(args[3])
+                limit = ast.literal_eval( args[4])
+                sort = ast.literal_eval(args[5])
+                search = ast.literal_eval(args[6])
+                res = manager.managers_ossec_log(type_log=type_log, category=category, months=months, offset=offset, limit=limit, sort=sort, search=search, cluster_depth=cluster_depth)
+            elif self.command[0] == list_requests_managers['MANAGERS_LOGS_SUMMARY']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                months = ast.literal_eval(args[0])
+                res = manager.managers_ossec_log_summary(months=months, cluster_depth=cluster_depth)
+            elif self.command[0] == list_requests_managers['MANAGERS_STATS_TOTALS']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                year = ast.literal_eval(args[0])
+                month = ast.literal_eval(args[1])
+                day = ast.literal_eval(args[2])
+                res = stats.totals(year=year, month=month, day=day, cluster_depth=cluster_depth)
+            elif self.command[0] == list_requests_managers['MANAGERS_STATS_HOURLY']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = stats.hourly(cluster_depth=cluster_depth)
+            elif self.command[0] == list_requests_managers['MANAGERS_STATS_WEEKLY']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = stats.weekly(cluster_depth=cluster_depth)
+            elif self.command[0] == list_requests_managers['MANAGERS_OSSEC_CONF']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                section = args[0]
+                field = ast.literal_eval(args[1])
+                res = manager.managers_get_ossec_conf(section=section, field=field, cluster_depth=cluster_depth)
+            elif self.command[0] == list_requests_managers['MANAGERS_INFO']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = myWazuh.managers_get_ossec_init(cluster_depth=cluster_depth)
+            elif self.command[0] == list_requests_cluster['CLUSTER_CONFIG']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                cluster_depth = ast.literal_eval(self.command[1]) - 1
+                res = get_config_distributed(cluster_depth=cluster_depth)
+
+            elif self.command[0] == list_requests_cluster['MASTER_FORW']:
+                args = self.f.decrypt(response[common.cluster_sync_msg_size:])
+                args = args.split(" ")
+                args_list = []
+                if args[0] in all_list_requests.values():
+                    agent_id = None
+                    request_type = args[0]
+                    if (len(args) > 1):
+                        args_list = args[1:]
+                elif len(args) > 1 and args[1] in all_list_requests.values():
+                    agent_id = args[0].split("-")
+                    request_type = args[1]
+                    if (len(args) > 2):
+                        args_list = args[2:]
+                res = distributed_api_request(request_type=request_type, agent_id=agent_id, args=args_list, cluster_depth=1, affected_nodes=None, from_cluster=True)
+
+            elif self.command[0] == list_requests_cluster['ready']:
                 res = "Starting to sync client's files"
                 # execute an independent process to "crontab" the sync interval
                 kill(child_pid, SIGUSR1)
+            elif self.command[0] == list_requests_cluster['data']:
+                res = "Saving data from actual master"
+                actual_master_data = json.loads(self.f.decrypt(response[common.cluster_sync_msg_size:]).decode())
+                save_actual_master_data_on_db(actual_master_data)
 
             logging.debug("Command {0} executed for {1}".format(self.command[0], self.addr))
 
@@ -138,7 +352,7 @@ class WazuhClusterHandler(asynchat.async_chat):
         else:
             error = str(v)
 
-        logging.error("Error handling client request: {0}".format(error))
+        logging.error("Error handling request: {0}".format(error))
         self.data = json.dumps({'error': 1, 'data': error})
         self.handle_write()
 
@@ -172,7 +386,7 @@ class WazuhClusterServer(asyncore.dispatcher):
         self.listen(50)
 
         cluster_info = read_config()
-        logging.info("Starting cluster {0}".format(cluster_info['name']))
+        logging.info("Starting cluster '{0}'".format(cluster_info['name']))
         logging.info("Listening on port {0}.".format(port))
         logging.info("{0} nodes found in configuration".format(len(cluster_info['nodes'])))
         logging.info("Synchronization interval: {0}".format(cluster_info['interval']))
@@ -198,23 +412,31 @@ def crontab_sync_master(interval):
     while True:
         logging.debug("Crontab: starting to sync")
         try:
-            sync(False)
+            sync_results = sync(True)
         except Exception as e:
             logging.error(str(e))
             kill(child_pid, SIGINT)
 
         config_cluster = read_config()
         for node in get_remote_nodes():
-            # ask clients to send updates
-            error, response = send_request(host=node, port=config_cluster["port"], key=config_cluster['key'],
-                                data="ready {0}".format('a'*(common.cluster_protocol_plain_size - len("ready "))))
+            if node[1] == 'master':
+                # send the synchronization results to the rest of masters
+                message = "data {0}".format('a'*(common.cluster_protocol_plain_size - len('data ')))
+                file = json.dumps(sync_results).encode()
+            else:
+                # ask clients to send updates
+                message = "ready {0}".format('a'*(common.cluster_protocol_plain_size - len("ready ")))
+                file = None
+
+            error, response = send_request(host=node[0], port=config_cluster["port"], key=config_cluster['key'],
+                                data=message, file=file)
 
         sleep(interval_number if interval_measure == 's' else interval_number*60)
 
 def crontab_sync_client():
     def sync_handler(n_signal, frame):
-        master = get_remote_nodes()[0]
-        sync_one_node(False, master)
+        master_ip, _, master_name = get_remote_nodes()[0]
+        sync_one_node(False, master_ip, master_name)
 
     signal(SIGUSR1, sync_handler)
     while True: