+++ framework/wazuh/__init__.py
@@ -6,6 +6,9 @@
 from wazuh import common
 from wazuh.utils import execute
 from wazuh.database import Connection
+from wazuh.cluster.distributed_api import is_a_local_request, distributed_api_request, is_cluster_running
+from wazuh.cluster.protocol_messages import list_requests_managers
+
 from time import strftime
 import re
 
@@ -130,6 +133,17 @@ class Wazuh:
             self.tz_name = None
 
         return self.to_dict()
+
+
+    def managers_get_ossec_init(self, node_id=None, cluster_depth=1):
+        if is_a_local_request() or cluster_depth <= 0 :
+            return self.get_ossec_init()
+        else:
+            if not is_cluster_running():
+                raise WazuhException(3015)
+
+            request_type = list_requests_managers['MANAGERS_INFO']
+            return distributed_api_request(request_type=request_type, cluster_depth=cluster_depth, affected_nodes=node_id)
 
 
 def main():